# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flask Demo",
    "description": "A simple demo app for Flask",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
      "Person": {
        "type": "object",
        "properties": {
          "job": {
            "type": "string",
            "description": "Job title"
          },
          "company": {
            "type": "string",
            "description": "Company name"
          },
          "ssn": {
            "type": "string",
            "description": "Social Security Number"
          },
          "residence": {
            "type": "string",
            "description": "Residence"
          },
          "current_location": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "blood_group": {
            "type": "string"
          },
          "website": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sex": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "mail": {
            "type": "string"
          },
          "birthdate": {
            "type": "datetime"
          }
        },
        "required": [
          "username",
          "name"
        ]
      },
      "SuggestionsResponse": {
        "type": "object",
        "properties": {
          "suggestions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of similar user names"
          }
        }
      },
      "PaginatedPersonsResponse": {
        "type": "object",
        "properties": {
          "persons": {
            "type": "array",
            "items": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Person"
                }
              ]
            },
            "description": "List of Persons"
          },
          "page": {
            "type": "integer"
          },
          "per_page": {
            "type": "integer"
          },
          "total_pages": {
            "type": "integer"
          },
          "prev_page": {
            "type": "string"
          },
          "next_page": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/search/{username}": {
      "get": {
        "summary": "Returns a user by username.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "Username to search"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/people": {
      "get": {
        "summary": "Returns a list of persons with pagination",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "default": 1,
            "description": "Page number for pagination"
          },
          {
            "name": "per_page",
            "in": "query",
            "type": "integer",
            "default": 10,
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPersonsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Creates and stores the given Person object in the backend.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        }
      }
    },
    "/people/{username}": {
      "delete": {
        "summary": "Deletes a user.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "Username to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted the person."
          },
          "404": {
            "description": "Person not found."
          }
        }
      },
      "put": {
        "summary": "Updates and stores the given Person object in the backend.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "type": "string",
            "required": true,
            "description": "Username to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Person successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        }
      }
    }
  }
}